Moved DP9800 modules to temperature directory
