+++ finesse/hardware/__init__.py
@@ -9,9 +9,9 @@ else:
     from .opus.em27 import OPUSInterface  # type: ignore
 
 if "--dummy-dp9800" in sys.argv:
-    from .dummy_dp9800 import DummyDP9800 as DP9800
+    from .temperature.dummy_dp9800 import DummyDP9800 as DP9800
 else:
-    from .dp9800 import DP9800  # type: ignore
+    from .temperature.dp9800 import DP9800  # type: ignore
 
 from .stepper_motor.dummy import DummyStepperMotor
