+++ finesse/hardware/temperature/dummy_dp9800.py
